reimplement utils file
